+++ crates/graphql-composition/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 - Composition now allows federated graphs without a query root. That happens if none of the subgraphs define a query root type. The new, relaxed requirement is that at least one root field (query, mutation or subscription) is defined, so the federated schema is not empty. (https://github.com/grafbase/grafbase/pull/3111)
+- Fixed a few inaccurate directive definition locations in the federated SDL emitted by render_federated_sdl().
 
 ### Fixes